made changes to whole game logic so it supports cell object. Everything works fine
