--- about.txt
@@ -1,12 +1,0 @@
-Labirynth ver. 1.1
-A simple, yet fun game where your goal is to exit the maze you're trapped in!
-The mazes created by this program are randomly generated through (yet) simple algorithms.
-From main menu you can access OPTIONS menu, where you can change your control settings
-and size of the maze. The default settings are:
-size - 5,
-up - 'u',
-down - 'd',
-left - 'l',
-right - 'r'.
-To exit the game type 'quit' or choose '4' in the main menu.
-Press ENTER to continue.
