+++ mvmnt_control.py
@@ -1,20 +1,20 @@
-from symbols import symbls
+SYMBOLS = '0123456789abcdefghijklmnopqrstuvwyz'
+from math import sqrt
 
-symbols, rows, columns = symbls()
-
-def move(pos, dirs, dir, grid, walls):
-	splt_pos = [char for char in pos]
+def move(pos, dirs, dir, grid):
+	i = SYMBOLS.index(pos[0])
+	j = SYMBOLS.index(pos[1])
+	size = int(sqrt(len(grid)))
 	if dir in dirs.values():
-		if dir == dirs['up'] and grid[symbols.index(splt_pos[0])*2+1][symbols.index(splt_pos[1])+1] != '--':
-			if pos == '00':
-				return 'win'
-			else:
-				splt_pos[0] = symbols[symbols.index(splt_pos[0])-1]
-		elif dir == dirs['down'] and grid[symbols.index(splt_pos[0])*2+3][symbols.index(splt_pos[1])+1] != '--':
-			splt_pos[0] = symbols[symbols.index(splt_pos[0])+1]
-		elif dir == dirs['left'] and symbols.index(splt_pos[1]) > 0 and walls[symbols.index(splt_pos[0])][symbols.index(splt_pos[1])-1] != 1:
-			splt_pos[1] = symbols[symbols.index(splt_pos[1])-1]
-		elif dir == dirs['right'] and symbols.index(splt_pos[1]) < len(walls)-1 and walls[symbols.index(splt_pos[0])][symbols.index(splt_pos[1])] != 1:
-			splt_pos[1] = symbols[symbols.index(splt_pos[1])+1]
-	pos = str(splt_pos[0] + splt_pos[1])
+		if dir == dirs['up'] and pos == '00':
+			return 'win'
+		elif dir == dirs['up'] and grid[i + j * size].walls["top"] != True:
+			j -= 1
+		elif dir == dirs['down'] and grid[i + j * size].walls["bottom"] != True:
+			j += 1
+		elif dir == dirs['left'] and grid[i + j * size].walls["left"] != True:
+			i -= 1
+		elif dir == dirs['right'] and grid[i + j * size].walls["right"] != True:
+			i += 1
+	pos = SYMBOLS[i] + SYMBOLS[j]
 	return pos
